Added a couple more tapering functions.
git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@6321 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ni/src/lib/nfp/rhombtriW.c
@@ -8,7 +8,7 @@
 
 extern void NGCALLF(drhombtrunc,DRHOMBTRUNC)(int*,int*,double*,double*,
                                              int*);
-extern void NGCALLF(dtritrunc,DTRITRUNC)(int*,int*,double*,double*,int*);
+extern void NGCALLF(dtritrunc,DTRITRUNC)(int*,int*,int*,double*,double*);
 
 NhlErrorTypes rhomb_trunC_W( void )
 {
